Import seller's catalog and store it in state
We need the square and dotenv packages for this.

listCatalog is a helper function that requests the seller's catalog from the Square API.

toObject accounts for the Express backend not being able to serialize BigInts when returning JSON.

The getMessage and addInventoryItem routes are request tests and will eventually be deleted.