Fix timezone error with order creation time
When the order creation timestamp was parsed in the backend, JS converted it to GMT+0 time, which was not helpful for users.

Fix this error by saving the RFC 3339 timestamp to the database. Parse it in the browser instead, which knows the user's timezone.

This should render the correct time of order creation.